db-crud/requirements.txt
mysql-connector-python==8.0.33
pytest==7.4.0
python-dotenv==1.0.0 